+++ backend/src/classes/Brave/Core/Migrations/Version20190210120518.php
@@ -1,0 +1,37 @@
+<?php declare(strict_types=1);
+
+namespace Brave\Core\Migrations;
+
+use Brave\Core\Entity\Role;
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Auto-generated Migration: Please modify to your needs!
+ */
+final class Version20190210120518 extends AbstractMigration
+{
+    /**
+     * @throws \Exception
+     */
+    public function up(Schema $schema) : void
+    {
+        // this up() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql',
+            'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('INSERT INTO roles (id, name) VALUES (12, "'.Role::APP_ESI.'")');
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function down(Schema $schema) : void
+    {
+        // this down() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql',
+            'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('DELETE FROM roles WHERE id = 12');
+    }
+}