+++ backend/src/tests/Functional/Core/Api/App/ApplicationControllerTest.php
@@ -7,6 +7,8 @@ use Brave\Core\Entity\Role;
 use Brave\Core\Entity\SystemVariable;
 use Brave\Core\Factory\RepositoryFactory;
 use Brave\Core\Entity\Group;
+use GuzzleHttp\ClientInterface;
+use Tests\Client;
 use Tests\WebTestCase;
 use Tests\Helper;
 use Brave\Core\Entity\Corporation;
@@ -49,7 +51,7 @@ class ApplicationControllerTest extends WebTestCase
         $this->helper->emptyDb();
         $group = (new Group())->setName('g1');
         $this->helper->getEm()->persist($group);
-        $app = $this->helper->addApp('Test App', 'boring-test-secret', ['app'])->addGroup($group);
+        $app = $this->helper->addApp('Test App', 'boring-test-secret', [Role::APP])->addGroup($group);
         $this->helper->getEm()->flush();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($app->getId().':boring-test-secret')];
@@ -73,7 +75,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testGroupsV1404()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
         $response = $this->runApp('GET', '/api/app/v1/groups/123', null, $headers);
@@ -84,7 +86,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testGroupsV2404()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
         $response = $this->runApp('GET', '/api/app/v2/groups/123', null, $headers);
@@ -218,7 +220,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testCorpGroupsV1404()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
         $response = $this->runApp('GET', '/api/app/v1/corp-groups/123', null, $headers);
@@ -229,7 +231,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testCorpGroupsV2404()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
         $response = $this->runApp('GET', '/api/app/v2/corp-groups/123', null, $headers);
@@ -300,7 +302,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testAllianceGroupsV1404()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
         $response = $this->runApp('GET', '/api/app/v1/alliance-groups/123', null, $headers);
@@ -311,7 +313,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testAllianceGroupsV2404()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
         $response = $this->runApp('GET', '/api/app/v2/alliance-groups/123', null, $headers);
@@ -382,7 +384,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testMainV1404()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
         $response = $this->runApp('GET', '/api/app/v1/main/123', null, $headers);
@@ -394,7 +396,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testMainV2404()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
         $response = $this->runApp('GET', '/api/app/v2/main/123', null, $headers);
@@ -406,7 +408,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testMainV1204()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $char = $this->helper->addCharacterMain('C1', 123, [Role::USER]);
         $char->setMain(false);
@@ -421,7 +423,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testMainV1200()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
         $char = $this->helper->addCharacterMain('C1', 123, [Role::USER]);
         $this->helper->addCharacterToPlayer('C2', 456, $char->getPlayer());
 
@@ -458,7 +460,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testCharactersV1404()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
 
         $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
         $response = $this->runApp('GET', '/api/app/v1/characters/123', null, $headers);
@@ -470,7 +472,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testCharactersV1200()
     {
         $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
         $char = $this->helper->addCharacterMain('C1', 123, [Role::USER]);
         $this->helper->addCharacterToPlayer('C2', 456, $char->getPlayer());
 
@@ -510,7 +512,7 @@ class ApplicationControllerTest extends WebTestCase
         $this->assertEquals(403, $response1->getStatusCode());
 
         $this->helper->emptyDb();
-        $appId = $this->helper->addApp('A1', 's1', ['app'])->getId();
+        $appId = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
         $headers = ['Authorization' => 'Bearer '.base64_encode($appId.':s1')];
 
         $response2 = $this->runApp('GET', '/api/app/v1/corporation/10/member-tracking', null, $headers);
@@ -523,7 +525,7 @@ class ApplicationControllerTest extends WebTestCase
     public function testMemberTrackingV1200()
     {
         $this->helper->emptyDb();
-        $appId = $this->helper->addApp('A1', 's1', ['app', 'app-tracking'])->getId();
+        $appId = $this->helper->addApp('A1', 's1', [Role::APP, Role::APP_TRACKING])->getId();
         $corp = (new Corporation())->setId(10)->setTicker('t1')->setName('corp 1');
         $member1 = (new CorporationMember())->setId(110)->setName('m1')->setCorporation($corp)
             ->setLogonDate(new \DateTime('now -5 days'));
@@ -557,6 +559,116 @@ class ApplicationControllerTest extends WebTestCase
         $this->assertSame(null, $result2[0]['startDate']);
     }
 
+    public function testEsiV1403()
+    {
+        $response1 = $this->runApp('GET', '/api/app/v1/esi');
+        $this->assertEquals(403, $response1->getStatusCode());
+
+        $this->helper->emptyDb();
+        $appId = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
+        $headers = ['Authorization' => 'Bearer '.base64_encode($appId.':s1')];
+
+        $response2 = $this->runApp('GET', '/api/app/v1/esi', null, $headers);
+        $this->assertEquals(403, $response2->getStatusCode());
+    }
+
+    public function testEsiV1400()
+    {
+        $this->helper->emptyDb();
+        $appId = $this->helper->addApp('A1', 's1', [Role::APP, Role::APP_ESI])->getId();
+        $headers = ['Authorization' => 'Bearer '.base64_encode($appId.':s1')];
+
+        $response1 = $this->runApp('GET', '/api/app/v1/esi', null, $headers);
+        $this->assertSame(400, $response1->getStatusCode());
+        $this->assertSame('Path cannot be empty.', $response1->getReasonPhrase());
+
+        $response2 = $this->runApp(
+            'GET',
+            '/api/app/v1/esi/latest/characters/96061222/stats/',
+            null,
+            ['Authorization' => 'Bearer '.base64_encode($appId.':s1')]
+        );
+        $this->assertSame(400, $response2->getStatusCode());
+        $this->assertSame(
+            'The datasource parameter cannot be empty, it must contain an EVE character IDs',
+            $response2->getReasonPhrase()
+        );
+
+        $response3 = $this->runApp(
+            'GET',
+            '/api/app/v1/esi/latest/characters/96061222/stats/?datasource=96061222',
+            null,
+            ['Authorization' => 'Bearer '.base64_encode($appId.':s1')]
+        );
+        $this->assertSame(400, $response3->getStatusCode());
+        $this->assertSame('Character not found.', $response3->getReasonPhrase());
+    }
+
+    public function testEsiV1200()
+    {
+        $this->helper->emptyDb();
+        $this->helper->addCharacterMain('C1', 123, [Role::USER]);
+        $appId = $this->helper->addApp('A1', 's1', [Role::APP, Role::APP_ESI])->getId();
+
+        $httpClient = new Client();
+        $httpClient->setResponse(new \GuzzleHttp\Psr7\Response(
+            200,
+            [
+                'Content-Type' => ['application/json; charset=UTF-8'],
+                'ETag' => ['686897696a7c876b7e'],
+                'Expires' => ['Sun, 10 Feb 2019 19:22:52 GMT'],
+                'X-Esi-Error-Limit-Remain' => [100],
+                'X-Esi-Error-Limit-Reset' => [60],
+                'X-Pages' => [3],
+                'warning' => ['199 - This route has an upgrade available'],
+            ],
+            '{"key": "value"}'
+        ));
+
+        $response = $this->runApp(
+            'GET',
+            '/api/app/v1/esi/v3/characters/96061222/assets/?page=1&datasource=123',
+            [],
+            [
+                'Authorization' => 'Bearer '.base64_encode($appId.':s1'),
+                'If-None-Match' => '686897696a7c876b7e'
+            ],
+            [ClientInterface::class => $httpClient]
+        );
+
+        $this->assertSame(200, $response->getStatusCode());
+        $this->assertSame('{"key": "value"}', $response->getBody()->__toString());
+        $this->assertSame([
+            'Content-Type' => ['application/json; charset=UTF-8'],
+            'ETag' => ['686897696a7c876b7e'],
+            'Expires' => ['Sun, 10 Feb 2019 19:22:52 GMT'],
+            'X-Esi-Error-Limit-Remain' => ['100'],
+            'X-Esi-Error-Limit-Reset' => ['60'],
+            'X-Pages' => ['3'],
+            'warning' => ['199 - This route has an upgrade available'],
+        ], $response->getHeaders());
+    }
+
+    public function testEsiV1200PathAsParameter()
+    {
+        $this->helper->emptyDb();
+        $this->helper->addCharacterMain('C1', 123, [Role::USER]);
+        $appId = $this->helper->addApp('A1', 's1', [Role::APP, Role::APP_ESI])->getId();
+
+        $httpClient = new Client();
+        $httpClient->setResponse(new \GuzzleHttp\Psr7\Response(200, [], '{"key": "value"}'));
+
+        $response = $this->runApp(
+            'GET',
+            '/api/app/v1/esi?path='. urlencode('/v3/characters/96061222/assets/') . '&page=1&datasource=123',
+            null,
+            ['Authorization' => 'Bearer '.base64_encode($appId.':s1')],
+            [ClientInterface::class => $httpClient]
+        );
+        $this->assertSame(200, $response->getStatusCode());
+        $this->assertSame('{"key": "value"}', $response->getBody()->__toString());
+    }
+
     private function setUpDb($invalidHours = 0)
     {
         $this->helper->emptyDb();
@@ -566,7 +678,7 @@ class ApplicationControllerTest extends WebTestCase
         $this->group1Id = $groups[1]->getId();
         $this->group4Id = $groups[4]->getId();
 
-        $app = $this->helper->addApp('A1', 's1', ['app']);
+        $app = $this->helper->addApp('A1', 's1', [Role::APP]);
         $app->addGroup($groups[0]);
         $app->addGroup($groups[1]);
         $app->addGroup($groups[4]);