+++ CHANGELOG.md
@@ -5,6 +5,7 @@ dd mm yyyy
 
 - Added application-api.json interface file that contains only the API for applications.
 - Added a script that creates a distribution.
+- Added ESI proxy endpoint for apps.
 
 ## 0.6.0